# src/docs/apispec.yaml
openapi: 3.1.0

info:
  title: Test Movie API
  version: 1.0.0
  description: REST API for Movie.

servers:
  - url: http://localhost:3000

tags:
  - name: Auth
  - name: Movies

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: test@mail.com
        password:
          type: string
          example: password123

    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: test@mail.com
        password:
          type: string
          example: password123

    CreateMovieRequest:
      type: object
      required:
        - title
        - director
        - releaseYear
        - poster
      properties:
        title:
          type: string
          example: Interstellar
        director:
          type: string
          example: Christopher Nolan
        releaseYear:
          type: integer
          example: 2014
        description:
          type: string
          example: Science fiction movie.
        poster:
          type: string
          format: binary

    UpdateMovieRequest:
      type: object
      properties:
        title:
          type: string
          example: Updated title
        director:
          type: string
          example: Updated director
        description:
          type: string
          example: Updated desc

  response:
    Auth:
      LoginSuccess:
        properties:
          message:
            type: string
            example: Authorization success
          data:
            type: object
            properties:
              id:
                type: integer
                example: 1
              name:
                type: string
                example: John Doe
              email:
                type: string
                example: test@mail.com
              accessToken:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MX0.ohHl1rvunW7yffTBSpHVJuWOp4JCOl0NCDK37oDca6M
      LoginFailed:
        properties:
          message:
            type: string
            example: incorrect email or password
          code:
            type: strin
            example: AUTH_INVALID_CREDENTIALS
      RegisterSuccess:
        properties:
          message:
            type: string
            example: user registered successfully
          data:
            type: object
            properties:
              id:
                type: integer
                example: 1
              name:
                type: string
                example: John Doe
              email:
                type: string
                example: test@mail.com
      ValidationError:
        properties:
          message:
            type: string
            example: validation error
          code:
            type: string
            example: VALIDATION_ERROR
          errors:
            type: array
            items:
              type: object
              properties:
                field:
                  type: string
                  example: name
                message:
                  type: string
                  example: name is required
      DuplicateEmail:
        properties:
          message:
            type: string
            example: duplicate email key
          code:
            type: string
            example: VALIDATION_ERROR
          errors:
            items:
              type: object
              properties:
                field:
                  type: string
                  example: email
                message:
                  type: string
                  example: email has been used by another user
    Movies:
      GetAllMovieSuccess:
        properties:
          message:
            type: string
            example: data retrieved successfully
          data:
            type: array
      GetMovieSuccess:
        properties:
          message:
            type: string
            example: data retrieved successfully
          data:
            type: object
            properties:
              title:
                type: string
                example: Interstellar
              director:
                type: string
                example: Christopher Nolan
              releaseYear:
                type: integer
                example: 2014
              description:
                type: string
                example: Science fiction movie.
              posterUrl:
                type: string
                example: https://res.cloudinary.com/du3o3jhep/image/upload/v1783103337/test/movies/qpwjlswdxpeca35bxvke.jpg
      UpdateMovieSuccess:
        properties:
          message:
            type: string
            example: movie updated successfully
          data:
            type: object
            properties:
              title:
                type: string
                example: Updated title
              director:
                type: string
                example: Updated director
              releaseYear:
                type: integer
                example: 2014
              description:
                type: string
                example: Updated desc
              posterUrl:
                type: string
                example: https://res.cloudinary.com/du3o3jhep/image/upload/v1783103337/test/movies/qpwjlswdxpeca35bxvke.jpg
      CreateMovieSuccess:
        properties:
          message:
            type: string
            example: movie created successfully
          data:
            type: object
            properties:
              title:
                type: string
                example: Interstellar
              director:
                type: string
                example: Christopher Nolan
              releaseYear:
                type: integer
                example: 2014
              description:
                type: string
                example: Science fiction movie.
              posterUrl:
                type: string
                example: https://res.cloudinary.com/du3o3jhep/image/upload/v1783103337/test/movies/qpwjlswdxpeca35bxvke.jpg
      DeleteMovieSuccess:
        properties:
          message:
            type: string
            example: movie deleted successfully
      ValidationError:
        properties:
          message:
            type: string
            example: validation error
          code:
            type: string
            example: VALIDATION_ERROR
          errors:
            type: array
            items:
              type: object
              properties:
                field:
                  type: string
                  example: title
                message:
                  type: string
                  example: title is required
      NotFound:
        properties:
          message:
            type: string
            example: data movie not found
          code:
            type: string
            example: RESOURCE_NOT_FOUND

paths:
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/response/Auth/LoginSuccess"
        "401":
          description: Login failed
          content:
            application/json:
              schema:
                $ref: "#/components/response/Auth/LoginFailed"

  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/response/Auth/RegisterSuccess"
        "422":
          description: User validation error
          content:
            application/json:
              schema:
                $ref: "#/components/response/Auth/ValidationError"
        "409":
          description: Duplicate Email
          content:
            application/json:
              schema:
                $ref: "#/components/response/Auth/DuplicateEmail"
  /api/movies:
    get:
      tags:
        - Movies
      summary: Get all movies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Movies retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/response/Movies/GetAllMovieSuccess"

    post:
      tags:
        - Movies
      summary: Create movie
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateMovieRequest"
      responses:
        "201":
          description: Movie created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/response/Movies/CreateMovieSuccess"
        "422":
          description: Movie validation error
          content:
            application/json:
              schema:
                $ref: "#/components/response/Movies/ValidationError"

  /api/movies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags:
        - Movies
      summary: Get movie by ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Movie retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/response/Movies/GetMovieSuccess"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/response/Movies/NotFound"

    put:
      tags:
        - Movies
      summary: Update movie
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMovieRequest"
      responses:
        "200":
          description: Movie updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/response/Movies/UpdateMovieSuccess"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/response/Movies/NotFound"

    delete:
      tags:
        - Movies
      summary: Delete movie
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Movie deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/response/Movies/DeleteMovieSuccess"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/response/Movies/NotFound"